deleted:    err_codes.h 	Replaced functionality with an enum in 'misc.h'. modified:   ieee.c modified:   ieee.h 	Changed 'bool feq(float,float)' to 'bool feq(double,double)' new file:   linear.c new file:   linear.h modified:   makefile 	Implemented linear solution. modified:   main.c modified:   misc.h modified:   parsing.c modified:   parsing.h 	Made parser return status object instead of exit on failure. modified:   quadratic.c 	Fixed order of quadratic solutions listed. modified:   testing.c modified:   testmain.c 	Modified slightly because of header dependencies. modified:   validate.c 	Made this look better.

+++ src/ieee.c
@@ -1,8 +1,8 @@
 #include "libraries.h"
 #include "misc.h"
 
-bool feq(double a,double b){
-	if(fabs(a-b)<0.0001){//Might have to change this constant
+bool feq(float a,float b){
+	if(fabsf(a-b)<0.0001){//Might have to change this constant
 		return true;
 	}
 	return false;
